Added simpler pagination

+++ src/main/java/com/example/blog/service/impl/PostServiceImpl.java
@@ -1,7 +1,6 @@
 package com.example.blog.service.impl;
 
 import com.example.blog.dto.PostDto;
-import com.example.blog.dto.RowDto;
 import com.example.blog.model.Post;
 import com.example.blog.model.User;
 import com.example.blog.repository.PostRepository;
@@ -11,16 +10,12 @@ import com.example.blog.service.mapper.PostDtoMapper;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
-import java.util.ArrayList;
 import java.util.List;
-import java.util.stream.Stream;
 
 @Service
 @RequiredArgsConstructor
@@ -29,8 +24,6 @@ public class PostServiceImpl implements PostService {
     private final PostRepository postRepository;
     private final PostDtoMapper postDtoMapper;
     private final UserService userService;
-    @Value("${app.view.number-of-elements-in-row}")
-    private Integer numberOfElementsInRow;
 
 
     @Override
@@ -41,10 +34,7 @@ public class PostServiceImpl implements PostService {
 
     @Override
     public Page<Post> findAllPaged(Pageable pageable) {
-//        List<PostDto> all = postRepository.findAll(pageable).stream().map(postDtoMapper).toList();
         Page<Post> all = postRepository.findAll(pageable);
-//        List<RowDto> rows = convertToRows(all);
-//        Page<PostDto> rowsPage = new PageImpl<>(all);
         return all;
     }
 
@@ -73,8 +63,13 @@ public class PostServiceImpl implements PostService {
     }
 
     @Override
-    public void deletePostById(Long id) {
-        postRepository.deleteById(id);
+    public boolean deletePostById(Long id, String currentUserUsername) {
+        PostDto postFromDb = this.findPostById(id);
+        if (postFromDb.getAuthorUsername().equals(currentUserUsername)) {
+            postRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 
 }
